birthday got updated to reactions confirmations too

+++ cogs/birthday.py
@@ -230,8 +230,33 @@ class Birthdays(commands.Cog):
 	async def delete(self, ctx):
 		results = await collection.find_one({"_id": ctx.author.id})
 		if results != None:
-			await collection.delete_one({"_id": ctx.author.id})
-			await ctx.send("Succesfully deleted your birthday from the list! {}".format(ctx.author.mention))
+			def check(reaction, user):
+				return str(reaction.emoji) in ['<:agree:797537027469082627>', '<:disagree:797537030980239411>'] and user.id == ctx.author.id
+			msg = await ctx.send("Are you sure you want to remove your birthday? %s" % (ctx.author.mention))
+			await msg.add_reaction('<:agree:797537027469082627>')
+			await msg.add_reaction('<:disagree:797537030980239411>')
+			try:
+				reaction, user = await self.client.wait_for('reaction_add', check=check, timeout=180)
+
+			except asyncio.TimeoutError:
+				new_msg = f"{ctx.author.mention} Did not react in time."
+				await msg.edit(content=new_msg)
+				await msg.clear_reactions()
+				return
+			
+			else:
+				if str(reaction.emoji) == '<:agree:797537027469082627>':
+					await collection.delete_one({"_id": ctx.author.id})
+					e = "Succesfully removed your birthday from the list! {}".format(ctx.author.mention)
+					await msg.edit(content=e)
+					await msg.clear_reactions()
+					return
+				
+				elif str(reaction.emoji) == '<:disagree:797537030980239411>':
+					e = "Birthday has not been removed. %s" % (ctx.author.mention)
+					await msg.edit(content=e)
+					await msg.clear_reactions()
+					return
 
 		else:
 			await ctx.send("You did not set your birthday, therefore you don't have what to delete! Type: `!birthday set <day | month>` to set your birthday.")
